Moved function generatenrandkeys to privatekey
