+++ internal/producer/src/producer/producer_test.go
@@ -8,9 +8,7 @@ import (
 	"database/sql"
 	"encoding/binary"
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"os"
 	"reflect"
@@ -951,145 +949,3 @@ func TestRecoverBlockchainMetadata_TwoBlocks(t *testing.T) {
 		})
 	}
 }
-
-func TestGenerateNRandomKeys(t *testing.T) {
-	type args struct {
-		filename string
-		n        uint32
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "Test with N = 0",
-			args: args{
-				filename: "testfile.json",
-				n:        0,
-			},
-			wantErr: true,
-			// Error should say "must generate at least 1 key"
-		},
-		{
-			name: "Test with N = 100",
-			args: args{
-				filename: "testfile.json",
-				n:        100,
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := GenerateNRandomKeys(tt.args.filename, tt.args.n)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("GenerateNRandomKeys() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			if _, err := os.Stat(tt.args.filename); os.IsNotExist(err) {
-				t.Errorf("Test file for keys not detected: %s", err)
-			}
-			if tt.args.n > 0 {
-				jsonFile, err := os.Open(tt.args.filename)
-				if err != nil {
-					t.Errorf("Failed to open json file: %s", err)
-				}
-				defer jsonFile.Close()
-				var keys []string
-				byteKeys, err := ioutil.ReadAll(jsonFile)
-				if err != nil {
-					t.Errorf("Failed to read in keys from json file: %s", err)
-				}
-				err = json.Unmarshal(byteKeys, &keys)
-				if err != nil {
-					t.Errorf("Failed to unmarshall keys because: %s", err)
-				}
-				if uint32(len(keys)) != tt.args.n {
-					t.Errorf("Number of private keys do not match n: %d", len(keys))
-				}
-
-			}
-		})
-	}
-	if err := os.Remove("testfile.json"); err != nil {
-		t.Errorf("Failed to remove file: %s because: %s", "testfile.json", err)
-	}
-}
-
-// func TestGenesisReadsAppropriately(t *testing.T) {
-// 	var testGenesisHash = "8db5d191bf333f96179c5f2ec7acd20a8c01378a1af120e2f2ded3672896931a"
-// 	genHashfile, _ := os.Create(constants.GenesisHashFile)
-// 	genHashfile.WriteString(testGenesisHash + "\n") // from GenerateGenesisHashFile
-// 	genHashfile.Close()
-// 	defer func() {
-// 		if err := os.Remove(constants.GenesisAddresses); err != nil {
-// 			t.Errorf("failed to remove genesis hash file: %v", err)
-// 		}
-// 	}()
-// 	hashSlice, err := genesis.ReadGenesisHashes()
-// 	if err != nil {
-// 		t.Errorf("failed to read from hash file: %v", hashSlice)
-// 	}
-// 	if hex.EncodeToString(hashSlice[0]) != testGenesisHash {
-// 		t.Errorf("hash from genesis_hashes.txt don't match: %s != %s", hex.EncodeToString(hashSlice[0]), testGenesisHash)
-// 	}
-// 	genesisBlock, err := genesis.BringOnTheGenesis(hashSlice, 1000)
-// 	if err != nil {
-// 		t.Errorf("failed to create genesis block: %v", err)
-// 	}
-// 	serializedCtc := genesisBlock.Data[0]
-// 	var ctc contracts.Contract
-// 	err = ctc.Deserialize(serializedCtc)
-// 	if err != nil {
-// 		t.Errorf("failed to deserialize block: %v", err)
-// 	}
-// 	recipient := hex.EncodeToString(ctc.RecipPubKeyHash)
-// 	if recipient != testGenesisHash {
-// 		t.Errorf("hashes don't match: %s != %s", recipient, testGenesisHash)
-// 	}
-
-// 	//airdrop
-// 	err = genesis.Airdrop("blockchain.dat", constants.MetadataTable, constants.AccountsTable, genesisBlock)
-// 	if err != nil {
-// 		t.Errorf("failed to airdrop genesis block: %s", err.Error())
-// 	}
-// 	defer func() {
-// 		if err := os.Remove("blockchain.dat"); err != nil {
-// 			t.Errorf("failed to remove blockchain.dat:\n%s", err.Error())
-// 		}
-// 		if err := os.Remove(constants.MetadataTable); err != nil {
-// 			t.Errorf("failed to remove metadatata.tab:\n%s", err.Error())
-// 		}
-// 		if err := os.Remove(constants.AccountsTable); err != nil {
-// 			t.Errorf("failed to remove accounts.db:\n%s", err.Error())
-// 		}
-// 	}()
-
-// 	db, err := sql.Open("sqlite3", constants.AccountsTable)
-// 	if err != nil {
-// 		t.Errorf("failed to open accounts table")
-// 	}
-// 	defer db.Close()
-
-// 	rows, err := db.Query(sqlstatements.GET_PUB_KEY_HASH_BALANCE_NONCE_FROM_ACCOUNT_BALANCES)
-// 	if err != nil {
-// 		t.Errorf("failed to create rows for queries")
-// 	}
-// 	defer rows.Close()
-
-// 	var pkhash string
-// 	var balance int
-// 	var nonce int
-// 	for rows.Next() {
-// 		rows.Scan(&pkhash, &balance, &nonce)
-// 		if pkhash != testGenesisHash {
-// 			t.Errorf("hash in accounts table doesn't match: %s != %s\n", pkhash, testGenesisHash)
-// 		}
-// 		if balance != 1000 {
-// 			t.Errorf("balance in accounts table doesn't match: %v != %v\n", balance, 1000)
-// 		}
-// 		if nonce != 0 {
-// 			t.Errorf("nonce in accounts table doesn't match: %v != %v\n", nonce, 0)
-// 		}
-// 	}
-// }
